fixing csv bug and adding specification
